% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{svg_groups_build}
\alias{svg_groups_build}
\title{Build up a larger images from the groups}
\usage{
svg_groups_build(svg, group_ids = svg_groups_list(svg), folder = ".",
  output_format = "png", attribute = "fragment",
  output_prefix = "split", node = "*")
}
\arguments{
\item{svg}{The svg}

\item{group_ids}{The order by which the groups will be added}

\item{folder}{The folder in which the output will be saved}

\item{output_format}{Whether to output the svg, png or pdf}

\item{attribute}{What attribute to look for, defaults to id}

\item{node}{What svg to look for, defaults to a group (which can include layers)}
}
\description{
Build up a larger images from the groups
}
