% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{svg_groups_list}
\alias{svg_groups_list}
\title{List the groups in an svg}
\usage{
svg_groups_list(svg, patterns = c("^(?!g[0-9]*).*$",
  "^(?!layer[0-9]*).*$"), node = "*", attr = "id")
}
\arguments{
\item{svg}{The svg to use}

\item{patterns}{One or more regular expression patterns, used for consecutive filtering}
}
\description{
List the groups in an svg
}
