% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{svg_groups_opacify}
\alias{svg_groups_opacify}
\title{Change the opacity of svg groups}
\usage{
svg_groups_opacify(svg, groups, output, folder = ".", verbose = FALSE,
  export = "--export-area-page", trim = FALSE, node = "svg:g",
  attribute = "id", output_format = "png")
}
\arguments{
\item{svg}{The svg}

\item{groups}{Tibble containing an id column (the id of the group) and an opacity column (double between 0 and 1)}

\item{output}{The output name, will be appended with .svg and/or .png}

\item{folder}{The folder in which the output will be saved}

\item{verbose}{Verbosity}

\item{export}{Inkscape argument for exporting}

\item{trim}{Whether to trim the png}

\item{node}{What svg to look for, defaults to a group (which can include layers)}

\item{attribute}{What attribute to look for, defaults to id}

\item{output}{Whether to output the svg, png or pdf}
}
\value{
The location of the png image
}
\description{
Change the opacity of svg groups
}
