% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{svg_groups_split}
\alias{svg_groups_split}
\title{Create smaller images from multiple groups in an svg}
\usage{
svg_groups_split(svg, group_ids = svg_groups_list(svg), folder = ".",
  output_format = "png")
}
\arguments{
\item{svg}{The svg}

\item{group_ids}{The identifiers of the groups which will all be transparent expect one}

\item{folder}{The folder in which the output will be saved}
}
\description{
Create smaller images from multiple groups in an svg
}
